Restructuring complete
